Finish iterasi 1

+++ routes/web.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\MemberController;
 use App\Http\Controllers\OwnerController;
 use App\Http\Controllers\RoomController;
 use App\Models\Booking;
+use illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -23,14 +24,14 @@ use Illuminate\Support\Facades\Route;
 //     return view('welcome');
 // });
 
-Route::resource('room', RoomController::class);
-Route::resource('member', MemberController::class);
+Route::resource('room', RoomController::class)->middleware('isAdminOwner');
+Route::resource('member', MemberController::class)->middleware('isAdminOwner');
 // Route::resource('booking', BookingController::class);
 
 // Route::get('/booking/{bookingg}', [BookingController::class, 'prevBooking']);
 Route::controller(BookingController::class)->group(function () {
-    Route::get('/show-room/{room}', 'prevBooking');
-    Route::post('/booking', 'booking');
+    Route::get('/show-room/{room}', 'prevBooking')->middleware('isMember');
+    Route::post('/booking', 'booking')->middleware('isMember');
 });
 
 Route::get('/test', function () {
@@ -53,20 +54,25 @@ Route::get('/test', function () {
     return $tabrakan;
 });
 
-Route::get('/admin', function () {
-    return view('Admin.Layouts.main');
-});
+Route::get('/dashboard', [OwnerController::class, 'dashboard'])->middleware('isAdminOwner');
 
 Route::get('/', function () {
+
+    if (Auth::guard('admin')->check() || Auth::guard('owner')->check()) {
+        return redirect('/dashboard');
+    }
+
     return view('Member.home');
 });
 
 Route::controller(RoomController::class)->group(function () {
-    Route::get('view-a', 'viewA');
+    Route::get('view-a', 'viewA')->middleware('isMember');
+    Route::get('view-b', 'viewB')->middleware('isMember');
+    Route::get('view-c', 'viewC')->middleware('isMember');
 });
 
 Route::get('/booking', function () {
-    return view('Member.booking');
+    return view('Member.booking')->middleware('isMember');
 });
 
 Route::get('/register', [MemberController::class, 'create']);
@@ -77,46 +83,33 @@ Route::controller(AuthController::class)->group(function () {
     Route::post('/register', 'register');
     Route::post('/logout', 'logout');
 });
+Route::get('detail/{booking}', [BookingController::class, 'detail'])->middleware('isMember');
 
-Route::get('/test1', function () {
-    return Auth::guard('admin')->user();
-});
+Route::get('/profile', [MemberController::class, 'profile'])->middleware('isMember');
 
-Route::get('/test2', function () {
-    return Auth::guard('member')->user();
-});
-
-Route::get('detail/{booking}', [BookingController::class, 'detail']);
-
-// Route::get('/profile', function () {
-//     return view('Member.profile');
-// });
-
-Route::get('/profile', [MemberController::class, 'profile']);
-
-Route::post('/upload-pembayaran/{booking}', [BookingController::class, 'uploadPembayaran']);
+Route::post('/upload-pembayaran/{booking}', [BookingController::class, 'uploadPembayaran'])->middleware('isMember');
 
 Route::controller(BookingController::class)->group(function () {
-    Route::get('/booking-admin', 'index');
-    Route::get('/booking-admin/{booking}', 'show');
-    Route::post('/set-status-pembayaran/{booking}', 'setStatusPembayaran');
-    Route::post('/set-status-pemesanan/{booking}', 'setStatusPemesanan');
-    Route::post('/laporan', 'laporan');
+    Route::get('/booking-admin', 'index')->middleware('isAdminOwner');
+    Route::get('/booking-admin/{booking}', 'show')->middleware('isAdminOwner');
+    Route::post('/set-status-pembayaran/{booking}', 'setStatusPembayaran')->middleware('isAdminOwner');
+    Route::post('/set-status-pemesanan/{booking}', 'setStatusPemesanan')->middleware('isAdminOwner');
+    Route::post('/laporan', 'laporan')->middleware('isAdminOwner');
 });
 
 Route::controller(MemberController::class)->group(function () {
-    Route::get('/member', 'index');
-    Route::get('/member/{member}/edit', 'edit');
-    Route::put('/member/{member}', 'update');
-    Route::delete('/member/{member}', 'destroy');
+    Route::get('/member', 'index')->middleware('isAdminOwner');
+    Route::get('/member/{member}/edit', 'edit')->middleware('isAdminOwner');
+    Route::put('/member/{member}', 'update')->middleware('isAdminOwner');
+    Route::delete('/member/{member}', 'destroy')->middleware('isAdminOwner');
 });
 
-Route::get('/edit-admin', [OwnerController::class, 'editAdmin']);
-Route::get('/edit-owner', [OwnerController::class, 'editOwner']);
+// Route::get('/edit-admin', [OwnerController::class, 'editAdmin'])->Middleware('isAdminOwner');
+// Route::get('/edit-owner', [OwnerController::class, 'editOwner'])->Middleware('isAdminOwner');
 
-Route::controller(OwnerController::class)->group(function() {
-    Route::get('/edit-admin', 'editAdmin');
-    Route::get('/edit-owner', 'editOwner');
-    Route::put('/update-owner/{owner}', 'updateOwner');
-    Route::put('/update-admin/{admin}', 'updateAdmin');
+Route::controller(OwnerController::class)->group(function () {
+    Route::get('/edit-admin', 'editAdmin')->Middleware('isAdminOwner');
+    Route::get('/edit-owner', 'editOwner')->Middleware('isAdminOwner');
+    Route::put('/update-owner/{owner}', 'updateOwner')->Middleware('isAdminOwner');
+    Route::put('/update-admin/{admin}', 'updateAdmin')->Middleware('isAdminOwner');
 });
